requirements.txt
absl-py==1.3.0
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.3.1
appdirs==1.4.4
asgiref==3.6.0
asn1crypto==0.24.0
astroid==2.5.3
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
bcrypt==4.0.1
bidict==0.22.1
billiard==3.6.4.0
CacheControl==0.12.11
cachetools==5.2.1
Cerberus==1.3.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloud-sql-python-connector
cloudpickle==2.2.0
colorama==0.4.4
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
cryptography==39.0.0
Cython
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.1
distro==1.8.0
Django==4.1.5
django-cors-headers
django-countries
django-crispy-forms==1.14.0
django-croppie==1.3
django-extensions==3.2.1
django-htmx==1.13.0
django-import-export==2.7.1
django-phone-field 
django-storages
django-timezone-field==4.2.3
django_debug_toolbar==3.8.1
django-widget-tweaks
djangorestframework
dnspython==1.16.0
docopt==0.6.2
email_validator
et-xmlfile==1.1.0
filelock==3.0.12
fire==0.5.0
flatbuffers==23.1.4
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
future==0.18.2
gast==0.5.3
greenlet==2.0.1
grpcio==1.51.1
google-cloud-storage
gunicorn==20.1.0
httptools==0.5.0
humanfriendly==10.0
idna==2.6
importlib-metadata==6.0.0
isort==5.8.0
jmespath==1.0.1
joblib==1.2.0
jsonpickle==2.2.0
jsonschema==3.2.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.4.4
kolo==2.3.0
lazy-object-proxy==1.6.0
libclang==15.0.6.1
locket==1.0.0
magic-filter==1.0.9
Markdown==3.4.1
MarkupPy==1.14
mccabe==0.6.1
msgpack==1.0.4
multidict==5.2.0
numpy==1.21.6
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==20.9
pamqp==3.2.1
pandas==1.3.5
paramiko==3.0.0
partd==1.3.0
pgdb==0.0.11
Pillow==9.4.0
pluggy==1.0.0
polars==0.18.0
prompt-toolkit==3.0.28
protobuf==3.19.6
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
PyJWT==2.6.0
pykwalify==1.8.0
pylint==2.7.4
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.7
python-dotenv==0.21.1
python-engineio==4.3.4
python-socketio==5.7.2
python3-openid==3.2.0
pyxdg==0.25
PyYAML
randomname==0.1.5
regex==2022.10.31
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
SecretStorage==2.3.1
sentry-sdk==1.11.1
setuptools
six==1.15.0
SQLAlchemy==1.4.46
sqlparse==0.4.1
tablib==3.3.0
tabulate==0.9.0
tarsafe==0.0.3
termcolor==2.2.0
terminaltables==3.1.10
texttable==1.6.7
threadpoolctl==3.1.0
tldextract==3.4.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tqdm==4.64.1
tstock==2.2.6
typeguard==2.13.3
typing-utils==0.1.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ujson==5.7.0
ulid-py==1.1.0
urllib3==1.26.13
uvloop==0.17.0
validate_email
vine==5.0.0
virtualenv==20.4.3
wcwidth==0.2.5
websocket-client==0.59.0
websockets==10.4
Werkzeug==2.0.3
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
yarl==1.8.2
zipp==3.11.0
